Address review feedback on BeanRegistry adapter seam
- Rename the new SPI to GrailsBeanRegistryAdapter (impl
  BeanBuilderGrailsBeanRegistryAdapter) to avoid colliding with Spring's
  org.springframework.beans.factory.support.BeanRegistryAdapter
- Validate the BeanBuilder constructor argument with Assert.notNull
- Align @since and the guide wording to 8.0.x and soften the #15824 reference
- Add the ASF license header to the new upgrading guide page
